src/main/java/com/ro/panel/player/PlayerService.java
package com.ro.panel.player;

import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class PlayerService {

    private final PlayerRepository playerRepository;

    public PlayerService(PlayerRepository playerRepository) {
        this.playerRepository = playerRepository;
    }

    public Iterable<Player> getAllPlayers(){
        return playerRepository.findAll();
    }

    public Optional<Player> getPlayerById(Integer id){
        return playerRepository.findById(id);
    }

    public void createPlayer(Player player){
        playerRepository.save(player);
    }
}
